Held item slot fix

+++ ac.artemis.packet.api/src/main/java/ac/artemis/packet/wrapper/server/PacketPlayServerItemHeldSlot.java
@@ -3,5 +3,5 @@ package ac.artemis.packet.wrapper.server;
 import ac.artemis.packet.wrapper.PacketServer;
 
 public interface PacketPlayServerItemHeldSlot extends PacketServer {
-    short getSlot();
+    byte getSlot();
 }
